refactor: create the basic file structure
